Issue: #14366
I have moved set arithmetic operators to sympy/sets/sets.py
from sympy/sets/setexpr.py
