+++ sympy/sets/setexpr.py
@@ -5,6 +5,7 @@ from sympy.sets import (imageset, Interval, FiniteSet, Union, ImageSet,
 from sympy.core.decorators import call_highest_priority, _sympifyit
 from sympy.utilities.iterables import sift
 from sympy.multipledispatch import dispatch, Dispatcher
+from sympy.sets import set_add, set_sub, set_mul, set_div, set_pow, set_function
 
 
 d = Dummy('d')
@@ -131,26 +132,3 @@ def _apply_operation(op, x, y, commutative):
             out = ImageSet(Lambda((_x, _y), op(_x, _y)), x, y)
     return out
 
-def set_add(x, y):
-    from sympy.sets.handlers.add import _set_add
-    return _apply_operation(_set_add, x, y, commutative=True)
-
-def set_sub(x, y):
-    from sympy.sets.handlers.add import _set_sub
-    return _apply_operation(_set_sub, x, y, commutative=False)
-
-def set_mul(x, y):
-    from sympy.sets.handlers.mul import _set_mul
-    return _apply_operation(_set_mul, x, y, commutative=True)
-
-def set_div(x, y):
-    from sympy.sets.handlers.mul import _set_div
-    return _apply_operation(_set_div, x, y, commutative=False)
-
-def set_pow(x, y):
-    from sympy.sets.handlers.power import _set_pow
-    return _apply_operation(_set_pow, x, y, commutative=False)
-
-def set_function(f, x):
-    from sympy.sets.handlers.functions import _set_function
-    return _set_function(f, x)
